substr

+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: aaudeber <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 12:05:27 by aaudeber          #+#    #+#             */
-/*   Updated: 2023/02/02 12:05:44 by aaudeber         ###   ########.fr       */
+/*   Updated: 2023/02/08 18:14:56 by aaudeber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
